+++ packages/components-js/tests/vrt/puppeteer/specs/tabs.vrt.ts
@@ -41,12 +41,9 @@ xit('should have no visual regression for :focus-visible', async () => {
           </p-tabs-item>
         </p-tabs>`;
 
-      await setContentWithDesignSystem(
-        page,
-        getThemedBodyMarkup(getElementsMarkup, { themes: ['light', 'dark', 'light-electric'], states: ['focus'] })
-      );
+      await setContentWithDesignSystem(page, getThemedBodyMarkup(getElementsMarkup, { states: ['focus'] }));
 
-      await forceFocusState(page, '.focus > p-tabs p-tabs-item');
+      await forceFocusState(page, '.focus p-tabs p-tabs-item');
     })
   ).toBeFalsy();
 });