+++ packages/components-react/tests/e2e/specs/jsdom-polyfill.e2e.ts
@@ -6,6 +6,6 @@
 import * as jsJsdomPolyfill from '@porsche-design-system/components-js/jsdom-polyfill';
 import * as reactJsdomPolyfill from '@porsche-design-system/components-react/jsdom-polyfill';
 
-it('should reexport jsdom-polyfill from components-js', () => {
+it('should reexport jsdom-polyfill sub-package from components-js', () => {
   expect(reactJsdomPolyfill).toEqual(jsJsdomPolyfill);
 });