+++ packages/components/src/components/accordion/accordion-styles.ts
@@ -1,7 +1,14 @@
 import type { BreakpointCustomizable, Theme } from '../../types';
 import type { AccordionSize } from './accordion-utils';
-import { buildResponsiveStyles, getCss } from '../../utils';
-import { getTransition, transitionDuration, getThemedColors, addImportantToEachRule } from '../../styles';
+import { buildResponsiveStyles, getCss, mergeDeep } from '../../utils';
+import {
+  getTransition,
+  transitionDuration,
+  getThemedColors,
+  addImportantToEachRule,
+  hostHiddenStyles,
+  hoverMediaQuery,
+} from '../../styles';
 import {
   fontWeight,
   fontSizeText,
@@ -11,8 +18,6 @@ import {
   borderRadiusSmall,
   borderWidthBase,
 } from '@porsche-design-system/utilities-v2';
-import { hoverMediaQuery } from '../../styles/hover-media-query';
-import { hostHiddenStyles } from '../../styles/host-hidden-styles';
 
 export const getComponentCss = (
   size: BreakpointCustomizable<AccordionSize>,
@@ -51,28 +56,35 @@ export const getComponentCss = (
           fontSize: fontSizeText[s],
           padding: compact ? '4px 0' : `${s === 'medium' ? '20px' : '12px'} 0`,
         })),
-        '&::before': {
-          content: '""',
-          position: 'absolute',
-          borderRadius: borderRadiusSmall,
-          left: '-4px',
-          right: '-4px',
-          ...(compact
-            ? {
-                top: '2px',
-                bottom: '2px',
-              }
-            : {
-                top: '6px',
-                bottom: '6px',
-              }),
-        },
-        ...hoverMediaQuery({
-          transition: getTransition('background-color'),
-          '&:hover::before': {
-            backgroundColor: hoverColor,
+        // mergeDeep needed because of hoverMediaQuery in certain modes not wrapping keys and therefore overriding "&::before" key
+        ...mergeDeep(
+          {
+            '&::before': {
+              content: '""',
+              position: 'absolute',
+              borderRadius: borderRadiusSmall,
+              left: '-4px',
+              right: '-4px',
+              ...(compact
+                ? {
+                    top: '2px',
+                    bottom: '2px',
+                  }
+                : {
+                    top: '6px',
+                    bottom: '6px',
+                  }),
+            },
           },
-        }),
+          hoverMediaQuery({
+            '&::before': {
+              transition: getTransition('background-color'),
+            },
+            '&:hover::before': {
+              background: hoverColor,
+            },
+          })
+        ),
         '&:focus::before': {
           border: `${borderWidthBase} solid ${focusColor}`,
         },