+++ packages/components/src/components/fieldset-wrapper/fieldset-wrapper-styles.spec.ts
@@ -1,14 +1,18 @@
 import { getComponentCss } from './fieldset-wrapper-styles';
 
-xdescribe('getComponentCss()', () => {
+describe('getComponentCss()', () => {
   it.each<Parameters<typeof getComponentCss>>([
-    ['none', 'medium', true],
-    ['none', 'medium', false],
-    ['none', 'small', true],
-    ['none', 'small', false],
-    ['success', 'medium', false],
-    ['error', 'medium', false],
-  ])('should return correct css for state: %s, labelSize: %s and hasLabel: %s', (...args) => {
+    ['none', 'medium', true, 'light'],
+    ['none', 'medium', true, 'dark'],
+    ['none', 'medium', false, 'light'],
+    ['none', 'small', true, 'light'],
+    ['none', 'small', true, 'dark'],
+    ['none', 'small', false, 'light'],
+    ['success', 'medium', false, 'light'],
+    ['success', 'medium', false, 'dark'],
+    ['error', 'medium', false, 'light'],
+    ['error', 'medium', false, 'dark'],
+  ])('should return correct css for state: %s, labelSize: %s, hasLabel: %s and theme: %s', (...args) => {
     expect(getComponentCss(...args)).toMatchSnapshot();
   });
 });