+++ packages/components/src/components/text/text-styles.spec.ts
@@ -1,10 +1,12 @@
-import { getComponentCss, getSlottedCss } from './text-styles';
+import { getComponentCss } from './text-styles';
 
-xdescribe('getComponentCss()', () => {
+describe('getComponentCss()', () => {
   it.each<Parameters<typeof getComponentCss>>([
+    ['small', 'regular', 'left', 'primary', false, 'light'],
     ['small', 'regular', 'left', 'default', false, 'light'],
     ['x-large', 'thin', 'right', 'brand', true, 'dark'],
     ['x-large', 'regular', 'center', 'brand', false, 'dark'],
+    ['large', 'semi-bold', 'left', 'notification-info', true, 'light'],
     ['large', 'semibold', 'left', 'notification-error', true, 'light'],
     ['medium', 'bold', 'right', 'neutral-contrast-high', true, 'dark'],
     ['x-small', 'semibold', 'left', 'inherit', false, 'light'],
@@ -23,16 +25,4 @@ xdescribe('getComponentCss()', () => {
       expect(getComponentCss(...args)).toMatchSnapshot();
     }
   );
-});
-
-xdescribe('getSlottedCss()', () => {
-  it('should return correct css', () => {
-    const host = document.createElement('p-text');
-    expect(getSlottedCss(host)).toMatchSnapshot();
-  });
-
-  it('should return correct css with prefix', () => {
-    const host = document.createElement('prefixed-p-text');
-    expect(getSlottedCss(host)).toMatchSnapshot();
-  });
 });