+++ packages/storefront/src/pages/patterns/typography/example-type-scale.vue
@@ -55,7 +55,7 @@
 </script>
 
 <style scoped lang="scss">
-  @import '~@porsche-design-system/components-js/utilities/scss';
+  @import '~@porsche-design-system/components-js/styles/scss';
 
   .type-scale {
     @include pds-grid;